# frontendmobileapp/swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "paths": {
...
        "/api/feed": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Fetch posts based on selected feed type (All Updates, Mentions, Favorites, Friends, Groups)",
                "operationId": "get_feed",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/FeedQuery"
                        }
                    }
                ],
                "tags": [
                    "api"
                ]
            }
        },
        "/api/follow/{keycloak_id}": {
            "parameters": [
                {
                    "name": "keycloak_id",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Follow or unfollow a user",
                "operationId": "post_follow_user",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/FollowUserRequest"
                        }
                    }
                ],
                "tags": [
                    "api"
                ]
            }
        },
        "/api/followers/{keycloak_id}": {
            "parameters": [
                {
                    "name": "keycloak_id",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Fetch all followers of a user",
                "operationId": "get_get_followers",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/GetFollowersRequest"
                        }
                    }
                ],
                "tags": [
                    "api"
                ]
            }
        },
        "/api/following/{keycloak_id}": {
            "parameters": [
                {
                    "name": "keycloak_id",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Fetch all users the current user is following",
                "operationId": "get_get_following",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/GetFollowingRequest"
                        }
                    }
                ],
                "tags": [
                    "api"
                ]
            }
        },
...
        "/api/post": {
            "post": {
                "responses": {
                    "201": {
                        "description": "Post created successfully"
                    }
                },
                "summary": "Create a new post",
                "operationId": "post_create_post",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/CreatePostRequest"
                        }
                    }
                ],
                "tags": [
                    "api"
                ]
            }
        },
        "/api/post/{post_id}/comment": {
            "parameters": [
                {
                    "name": "post_id",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Add a comment to a post",
                "operationId": "post_add_comment",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/AddCommentRequest"
                        }
                    }
                ],
                "tags": [
                    "api"
                ]
            }
        },
        "/api/post/{post_id}/comments": {
            "parameters": [
                {
                    "name": "post_id",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Fetch all comments for a post",
                "operationId": "get_get_comments",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/GetMessagesRequest"
                        }
                    }
                ],
                "tags": [
                    "api"
                ]
            }
        },
        "/api/post/{post_id}/like": {
            "parameters": [
                {
                    "name": "post_id",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Like or unlike a post",
                "operationId": "post_like_post",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/LikePostRequest"
                        }
                    }
                ],
                "tags": [
                    "api"
                ]
            }
        },
        "/api/post/{post_id}/reaction": {
            "parameters": [
                {
                    "name": "post_id",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Add or update a reaction to a post (like, love, laugh, etc",
                "description": ").",
                "operationId": "post_react_to_post",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/ReactionRequest"
                        }
                    }
                ],
                "tags": [
                    "api"
                ]
            }
        },
...
    },
    "info": {
        "title": "YesLove API",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
...
    "definitions": {
...
        "FeedQuery": {
            "properties": {
                "feed_type": {
                    "type": "string",
                    "description": "Type of feed: 'all', 'mentions', 'favorites', 'friends', 'groups'"
                }
            },
            "type": "object"
        },
        "CreatePostRequest": {
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "description": "Content of the post"
                }
            },
            "type": "object"
        },
        "ReactionRequest": {
            "required": [
                "reaction_type"
            ],
            "properties": {
                "reaction_type": {
                    "type": "string",
                    "description": "Reaction type: 'like', 'love', 'laugh', 'angry', etc."
                }
            },
            "type": "object"
        },
        "GetFollowersRequest": {
            "properties": {},
            "type": "object"
        },
        "LikePostRequest": {
            "required": [
                "post_id"
            ],
            "properties": {
                "post_id": {
                    "type": "integer",
                    "description": "ID of the post to like or unlike"
                }
            },
            "type": "object"
        },
        "AddCommentRequest": {
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "description": "Content of the comment"
                }
            },
            "type": "object"
        },
        "GetMessagesRequest": {
            "properties": {},
            "type": "object"
        },
        "FollowUserRequest": {
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "type": "string",
                    "description": "'follow' to follow, 'unfollow' to unfollow"
                }
            },
            "type": "object"
        },
        "GetFollowingRequest": {
            "properties": {},
            "type": "object"
        },
...
    },
...
}